write a selector to group transactions
